1) Added validations 2) Bugs fixed

+++ src/main/java/com/yegorf/controllers/repos/SymptomeRepo.java
@@ -5,6 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SymptomeRepo extends CrudRepository<Symptome, Long> {
-
+public interface SymptomeRepo extends CrudRepository<Symptome, Integer> {
 }
